Revert "Nhsdkk move models to separate folder"
