+++ clean/clean.go
@@ -1,10 +1,10 @@
 package clean
 
 import (
-	"GoRelCli/models/schema_model"
-	"GoRelCli/utils/logger"
-	"GoRelCli/utils/schema_parser"
-	"GoRelCli/utils/validator"
+	"GoRelCli/logger"
+	"GoRelCli/schema_model"
+	"GoRelCli/schema_parser"
+	"GoRelCli/validator"
 	"errors"
 	"gopkg.in/yaml.v3"
 	"os"
