+++ generate/generate.go
@@ -1,10 +1,10 @@
 package generate
 
 import (
-	"GoRelCli/models/schema_model"
-	"GoRelCli/utils/logger"
-	"GoRelCli/utils/schema_parser"
-	"GoRelCli/utils/validator"
+	"GoRelCli/logger"
+	"GoRelCli/schema_model"
+	"GoRelCli/schema_parser"
+	"GoRelCli/validator"
 	"errors"
 	"fmt"
 	"os"
