+++ gorel/gorel_schema.yml
@@ -1,45 +1,3 @@
-connection:
-  provider: postgresql
-  url: env("DATABASE_URL")
-#  url: "postgresql://postgres:POSTGRESPW@localhost:5432/postgres?sslmode=disable"
-
-models:
-  - name: User
-    properties:
-      - name: id
-        type: int
-        default: autoincrement()
-        id: true
-      - name: email
-        type: string
-        unique: true
-      - name: username
-        type: string?
-      - name: isVerified
-        type: boolean
-        default: false
-      - name: userType
-        type: UserRole
-      - name: todos
-        type: Todo[]
-
-  - name: Todo
-    properties:
-      - name: id
-        type: string
-        default: uuid()
-        id: true
-      - name: title
-        type: string
-      - name: userId
-        type: int
-      - name: user
-        type: User
-        relationField: userId
-        referenceField: id
-
-enums:
-  - name: UserRole
-    values:
-      - Admin
-      - User
+connection: {provider: postgresql, url: 'postgresql://postgres:POSTGRESPW@localhost:5432/postgres?sslmode=disable'}
+models: [{name: User, properties: [{name: id, type: int, default: autoincrement(), unique: false, id: true}, {name: email, type: string, unique: true, id: false}, {name: username, type: 'string?', unique: false, id: false}, {name: isVerified, type: boolean, default: "false", unique: false, id: false}, {name: userType, type: UserRole, unique: false, id: false}, {name: todos, type: 'Todo[]', unique: false, id: false}]}, {name: Todo, properties: [{name: id, type: string, default: uuid(), unique: false, id: true}, {name: title, type: string, unique: false, id: false}, {name: userId, type: int, unique: false, id: false}, {name: user, type: User, unique: false, id: false, relationField: userId, referenceField: id}]}]
+enums: [{name: UserRole, values: [Admin, User]}]
