+++ gorel_schema.yml
@@ -1,0 +1,45 @@
+connection:
+  provider: postgresql
+#  url: env("DATABASE_URL")
+  url: "postgresql://postgres:POSTGRESPW@localhost:5432/postgres?sslmode=disable"
+
+models:
+  - name: User
+    properties:
+      - name: id
+        type: int
+        default: autoincrement()
+        id: true
+      - name: email
+        type: string
+        unique: true
+      - name: username
+        type: string?
+      - name: isVerified
+        type: boolean
+        default: false
+      - name: userType
+        type: UserRole
+      - name: todos
+        type: Todo[]
+
+  - name: Todo
+    properties:
+      - name: id
+        type: string
+        default: uuid()
+        id: true
+      - name: title
+        type: string
+      - name: userId
+        type: int
+      - name: user
+        type: User
+        relationField: userId
+        referenceField: id
+
+enums:
+  - name: UserRole
+    values:
+      - Admin
+      - User
