+++ raw_sql/create_enums.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+CREATE TYPE userRole AS ENUM('Admin','User');
+CREATE TYPE userType AS ENUM('AdminType','UserType');
+
+COMMIT;
