+++ raw_sql/drop_enums.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP TYPE userrole;
+-- DROP TYPE usertype;
+
+END;
