+++ raw_sql/drop_schema.sql
@@ -1,0 +1,4 @@
+BEGIN;
+DROP SCHEMA public CASCADE ;
+CREATE SCHEMA public;
+COMMIT;
