+++ raw_sql/drop_tables.sql
@@ -1,0 +1,5 @@
+BEGIN;
+-- ALTER TABLE todo DROP CONSTRAINT fk_person;
+DROP TABLE IF EXISTS person CASCADE ;
+DROP TABLE IF EXISTS todo CASCADE ;
+END;
