+++ raw_sql/read_enums.sql
@@ -1,0 +1,4 @@
+select *
+FROM pg_type
+WHERE OID = ANY (select enumtypid
+FROM pg_enum);
