+++ raw_sql/read_schema.sql
@@ -1,0 +1,3 @@
+SELECT table_name
+FROM information_schema.tables
+WHERE table_schema = 'public';
